Refactor help message handling and add member/team list retrieval functions
