+++ libs/functions/compose/text_item.py
@@ -4,6 +4,8 @@ libs/functions/compose/text_item.py
 
 from typing import TYPE_CHECKING, Literal, Optional
 
+from table2ascii import Alignment, PresetStyle, table2ascii
+
 import libs.global_value as g
 from libs.utils.timekit import Delimiter, Format
 from libs.utils.timekit import ExtendedDatetime as ExtDt
@@ -200,3 +202,54 @@ def date_range(
             ret = f"{str_st} - {str_et}"
 
     return ret
+
+
+def get_members_list() -> str:
+    """登録済みのメンバー一覧を取得する
+
+    Returns:
+        str: メンバーリスト
+    """
+
+    name_list: list = []
+    for pname in g.cfg.member.lists:
+        name_list.append([pname, ", ".join(g.cfg.member.alias(pname))])
+
+    if name_list:
+        output = table2ascii(
+            header=["表示名", "登録されている名前"],
+            body=name_list,
+            alignments=[Alignment.LEFT, Alignment.LEFT],
+            style=PresetStyle.ascii_borderless,
+        )
+    else:
+        output = "メンバーは登録されていません。"
+
+    return output
+
+
+def get_team_list() -> str:
+    """チームの登録状況を取得する
+
+    Returns:
+        str: チームリスト
+    """
+
+    team_list: list = []
+    for team_name in g.cfg.team.lists:
+        if member := ", ".join(g.cfg.team.member(team_name)):
+            team_list.append([team_name, member])
+        else:
+            team_list.append([team_name, "未エントリー"])
+
+    if team_list:
+        output = table2ascii(
+            header=["チーム名", "所属メンバー"],
+            body=team_list,
+            alignments=[Alignment.LEFT, Alignment.LEFT],
+            style=PresetStyle.ascii_borderless,
+        )
+    else:
+        output = "チームは登録されていません。"
+
+    return output
